Index

This page lists the core methods and types of the NumericalTypeAliases.jl package. The Types section contains all type aliases, abstract and concrete. The Constants section also has some constant utility variables associated with the package. Each of these entries link to the docstrings in the Docs section.

Types

NumericalTypeAliases.Float

NumericalTypeAliases.IntegerArray

NumericalTypeAliases.IntegerMatrix

NumericalTypeAliases.IntegerVector

NumericalTypeAliases.RealArray

NumericalTypeAliases.RealFP

NumericalTypeAliases.RealMatrix

NumericalTypeAliases.RealVector

Constants

NumericalTypeAliases.NTA_ABSTRACT_TYPES

NumericalTypeAliases.NTA_CONCRETE_TYPES

NumericalTypeAliases.NTA_VERSION

Docs

NumericalTypeAliases — Module

Main module for NumericalTypeAliases.jl, a Julia package of numerical array type aliases.

This module exports all of the type aliases and utilities used by the NumericalTypeAliases.jl package.

Exports

Float

IntegerArray

IntegerMatrix

IntegerVector

NTA_ABSTRACT_TYPES

NTA_CONCRETE_TYPES

NTA_VERSION

RealArray

RealFP

RealMatrix

RealVector

NumericalTypeAliases.RealArray — Type

RealArray{T<:Real, N} = AbstractArray{T, N}

Abstract type for N-dimensional array of real, non-integered values.

NumericalTypeAliases.RealVector — Type

RealVector{T<:Real} = AbstractArray{T, 1}

Abstract type for 1-D vector of real, non-integered values.

NumericalTypeAliases.RealMatrix — Type

RealMatrix{T<:Real} = AbstractArray{T, 2}

Abstract type for 2-D matrix of real, non-integered values.

NumericalTypeAliases.IntegerArray — Type

IntegerArray{T<:Integer, N} = AbstractArray{T, N}

Abstract type for N-dimensional array of whole-numbered integers.

NumericalTypeAliases.IntegerVector — Type

IntegerVector{T<:Integer} = AbstractArray{T, 1}

Abstract type for 1-D vector of whole-numbered integers.

NumericalTypeAliases.IntegerMatrix — Type

IntegerMatrix{T<:Integer} = AbstractArray{T, 2}

Abstract type for 2-D matrix of whole-numbered integers.

NumericalTypeAliases.RealFP — Type

RealFP = AbstractFloat

Abstract type for a real floating-point number. This definition is provided for naming consistency in the package despite being equivalent to an AbstractFloat.

NumericalTypeAliases.Float — Type

Float

Concrete type for the default available floating point value. This is likely Float64 on most systems.

Because a Float64 in Julia is the equivalent of a double in other languages (while Float32 is a float), this has the possibility to be confusing depending on the context. Use carefully.

NumericalTypeAliases.NTA_VERSION — Constant

NTA_VERSION

A constant that contains the version of the installed NumericalTypeAliases.jl package.

NumericalTypeAliases.NTA_ABSTRACT_TYPES — Constant

NTA_ABSTRACT_TYPES

A list of NumericalTypeAliases' abstract types.

NumericalTypeAliases.NTA_CONCRETE_TYPES — Constant

NTA_CONCRETE_TYPES

A list of NumericalTypeAliases' concrete types.