Index

This page lists the core methods and types of the NumericalTypeAliases.jl package. The Modules section lists the modules exported by the package including the NumericalTypeAliases module itself. The Types section contains all type aliases, abstract and concrete. The Constants section also has some constant utility variables associated with the package. Each of these entries link to the docstrings in the Docs section.

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

NumericalTypeAliases.NumericalTypeAliases

Types

NumericalTypeAliases.Float

NumericalTypeAliases.IntegerArray

NumericalTypeAliases.IntegerMatrix

NumericalTypeAliases.IntegerVector

NumericalTypeAliases.RealArray

NumericalTypeAliases.RealFP

NumericalTypeAliases.RealMatrix

NumericalTypeAliases.RealVector

Constants

NumericalTypeAliases.NTA_ABSTRACT_TYPES

NumericalTypeAliases.NTA_CONCRETE_TYPES

NumericalTypeAliases.NTA_TYPES

NumericalTypeAliases.NTA_VERSION

Docs

This section lists the documentation for every exported name of the NumericalTypeAliases.jl package.

Modules

NumericalTypeAliases.NumericalTypeAliases — Module

Main module for NumericalTypeAliases.jl, a Julia package of numerical array type aliases.

This module exports all of the type aliases and utilities used by the NumericalTypeAliases.jl package.

Basic Usage

Install and import the package interactively with

julia> ]
(@v1.8) pkg> add NumericalTypeAliases

Then develop your package with numerical aliases such as

using NumericalTypeAliases

function do_something(data::RealMatrix, labels::IntegerVector)
    # Write a function that requires a 2-D float matrix and integer vector.
end

Imports

The following names are imported by the package as dependencies:

Base

Core

DocStringExtensions

Pkg

Exports

The following names are exported and available when using the package:

Float

IntegerArray

IntegerMatrix

IntegerVector

NTA_ABSTRACT_TYPES

NTA_CONCRETE_TYPES

NTA_TYPES

NTA_VERSION

RealArray

RealFP

RealMatrix

RealVector

Types

NumericalTypeAliases.Float — Type

Float

Description

Concrete type for the default available floating point value. This is likely Float64 on most systems.

Because a Float64 in Julia is the equivalent of a double in other languages (while Float32 is a float), this has the possibility to be confusing depending on the context. Use carefully.

NumericalTypeAliases.IntegerArray — Type

IntegerArray

Description

Abstract type for N-dimensional array of whole-numbered integers.

NumericalTypeAliases.IntegerMatrix — Type

IntegerMatrix

Description

Abstract type for 2-D matrix of whole-numbered integers.

NumericalTypeAliases.IntegerVector — Type

IntegerVector

Description

Abstract type for 1-D vector of whole-numbered integers.

NumericalTypeAliases.RealArray — Type

RealArray

Description

Abstract type for N-dimensional array of real, non-integered values.

NumericalTypeAliases.RealFP — Type

RealFP

Description

Abstract type for a real floating-point number. This definition is provided for naming consistency in the package despite being equivalent to an AbstractFloat.

NumericalTypeAliases.RealMatrix — Type

RealMatrix

Description

Abstract type for 2-D matrix of real, non-integered values.

NumericalTypeAliases.RealVector — Type

RealVector

Description

Abstract type for 1-D vector of real, non-integered values.

Constants

NumericalTypeAliases.NTA_ABSTRACT_TYPES — Constant

NTAABSTRACTTYPES

Description

A list of NumericalTypeAliases` abstract types.

NumericalTypeAliases.NTA_CONCRETE_TYPES — Constant

NTACONCRETETYPES

Description

A list of NumericalTypeAliases concrete types.

NumericalTypeAliases.NTA_TYPES — Constant

NTA_TYPES

Description

A combined list of abstract and concrete types exported in the NumericalTypeAliases package.

NumericalTypeAliases.NTA_VERSION — Constant

NTA_VERSION

Description

A constant that contains the version of the installed NumericalTypeAliases.jl package.

This value is computed at compile time, so it may be used to programmatically verify the version of NumericalTypeAliases that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.